# lambdas/auth-handler/docs/openapi.yaml
openapi: "3.0.3"

info:
  title: parking-app — Auth Handler API
  version: "3.0.0"
  description: |
    Endpoints de autenticación, registro y gestión de sesiones.

    Implementados como una sola Lambda Go (`auth-handler`) detrás de API Gateway HTTP v2.
    El identificador de usuario es el **RUT chileno** (formato `12345678-9`).

    Al registrar se escribe simultáneamente en **Cognito** (fuente de verdad de auth)
    y en **DynamoDB** (fuente de verdad de datos de negocio). Si DynamoDB falla,
    se hace rollback eliminando el usuario de Cognito.

    El login de operadores (`CUSTOMER_OPERATOR`) requiere adicionalmente el
    `serial_number` del terminal desde donde se inicia sesión. Cada login exitoso
    registra un ítem `SESSION#ACTIVE` en DynamoDB (TTL 24h) para enforcement de
    sesión única y soporte de logout remoto.

servers:
  - url: https://qnehzrs7g4.execute-api.us-east-1.amazonaws.com
    description: dev

security: []

components:
  securitySchemes:
    CognitoJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access token JWT emitido por Cognito.
        API Gateway valida la firma antes de invocar la Lambda.
        Enviar como `Authorization: Bearer <token>`.

  schemas:

    LoginRequest:
      type: object
      required: [rut, password]
      properties:
        rut:
          type: string
          description: |
            RUT chileno. Formatos aceptados: `12345678-9`, `12.345.678-9`, `123456789`.
            La Lambda normaliza internamente a `CUERPO-DV`.
          example: "12345678-5"
        password:
          type: string
          format: password
          minLength: 8
          example: "MiPass123!"
        serial_number:
          type: string
          description: |
            Serial del terminal TUU. **Requerido para usuarios CUSTOMER_OPERATOR.**
            Ignorado para ADMIN y CUSTOMER. La Lambda valida contra DynamoDB GSI1
            que el terminal esté registrado y no esté INACTIVE.
          example: "TUU-2024-001"

    LoginResponse:
      type: object
      required: [access_token, id_token, refresh_token, expires_in]
      properties:
        access_token:
          type: string
          description: JWT para llamadas a endpoints protegidos (`Authorization Bearer`).
        id_token:
          type: string
          description: JWT con atributos del usuario (email, nombre, `cognito:groups`).
        refresh_token:
          type: string
          description: Token de larga duración para renovar el `access_token` sin re-login.
        expires_in:
          type: integer
          description: Segundos hasta que expire el `access_token`.
          example: 3600

    RegisterRequest:
      type: object
      required: [rut, password, email, given_name, family_name]
      properties:
        rut:
          type: string
          description: RUT chileno (mismo formato que login).
          example: "12345678-5"
        password:
          type: string
          format: password
          minLength: 8
          description: Mínimo 8 caracteres, al menos una mayúscula y un número (política Cognito).
          example: "MiPass123!"
        email:
          type: string
          format: email
          example: "juan.perez@empresa.cl"
        given_name:
          type: string
          example: "Juan"
        family_name:
          type: string
          example: "Pérez"
        phone_number:
          type: string
          description: Teléfono en formato E.164. Opcional.
          pattern: "^\\+[1-9]\\d{1,14}$"
          example: "+56912345678"
        role:
          type: string
          description: Grupo Cognito a asignar. Por defecto `CUSTOMER_OPERATOR`.
          enum: [ADMIN, CUSTOMER, CUSTOMER_OPERATOR]
          default: CUSTOMER_OPERATOR
        org_id:
          type: string
          format: uuid
          description: |
            UUID de la organización a la que pertenece el operador.
            Requerido cuando `role = CUSTOMER_OPERATOR`.
          example: "c0c0c0c0-0000-0000-0000-000000000000"
        location_id:
          type: string
          format: uuid
          description: UUID de la sede asignada al operador. Opcional.
          example: "l1l1l1l1-0000-0000-0000-000000000000"
        org_name:
          type: string
          description: |
            Nombre de la organización. **Requerido cuando `role = CUSTOMER`.**
            El ADMIN crea la organización y el usuario CUSTOMER en una sola operación atómica.
          example: "Estacionamiento Central SpA"
        org_rut:
          type: string
          description: RUT de la empresa. Opcional, solo aplica cuando `role = CUSTOMER`.
          example: "76.543.210-K"

    RegisterResponse:
      type: object
      required: [message, id]
      properties:
        message:
          type: string
          example: "usuario_creado"
        id:
          type: string
          format: uuid
          description: UUID interno del usuario en DynamoDB.
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        org_id:
          type: string
          format: uuid
          description: UUID de la organización creada. Solo presente cuando `role = CUSTOMER`.
          example: "c0c0c0c0-0000-0000-0000-000000000000"

    BatchRegisterResponse:
      type: object
      required: [total, created, failed]
      properties:
        total:
          type: integer
          description: Total de filas de datos en el Excel (sin contar el header).
          example: 10
        created:
          type: integer
          example: 9
        failed:
          type: integer
          example: 1
        errors:
          type: array
          description: Detalle de cada fila fallida. Ausente si `failed` es 0.
          items:
            $ref: "#/components/schemas/BatchError"

    BatchError:
      type: object
      required: [row, rut, error]
      properties:
        row:
          type: integer
          description: Número de fila en el Excel (fila 1 es header, datos desde fila 2).
          example: 4
        rut:
          type: string
          example: "12345678-9"
        error:
          type: string
          enum: [rut_invalido, fila_incompleta, error_al_crear_usuario, error_dynamo]

    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: "sesion_cerrada"

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Código de error en snake_case.
          example: "error_interno"

tags:
  - name: Auth
    description: Login y registro (endpoints públicos)
  - name: Sessions
    description: Gestión de sesiones activas (requieren JWT)

paths:

  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Iniciar sesión
      description: |
        Autentica un usuario con RUT y contraseña contra Cognito.

        Tras autenticación exitosa:
        1. Decodifica el JWT para leer `cognito:groups` y `sub`.
        2. Si es `CUSTOMER_OPERATOR`: valida `serial_number` contra DynamoDB GSI1.
        3. Registra `USER#<id>/SESSION#ACTIVE` en DynamoDB (TTL 24h, best-effort).
        4. Retorna los tres tokens de Cognito.

        El `access_token` debe enviarse en `Authorization: Bearer <token>` para
        endpoints protegidos. Expira en ~1h. Usar `refresh_token` para renovar.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            examples:
              web_admin:
                summary: Login web (ADMIN / CUSTOMER)
                value:
                  rut: "12345678-5"
                  password: "MiPass123!"
              terminal_operator:
                summary: Login desde terminal (CUSTOMER_OPERATOR)
                value:
                  rut: "87654321-K"
                  password: "OtraPass456!"
                  serial_number: "TUU-2024-001"
      responses:
        "200":
          description: Autenticación exitosa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Body inválido, RUT incorrecto, o `serial_number` faltante para operador
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                serial_requerido:
                  value: { error: "serial_number_requerido" }
                rut_invalido:
                  value: { error: "rut_invalido" }
        "401":
          description: Contraseña incorrecta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "credenciales_invalidas"
        "403":
          description: Terminal no registrado o desactivado (solo CUSTOMER_OPERATOR)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                terminal_no_registrado:
                  value: { error: "terminal_no_registrado" }
                terminal_inactivo:
                  value: { error: "terminal_inactivo" }
        "404":
          description: Usuario no encontrado en Cognito
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "usuario_no_encontrado"
        "500":
          description: Error interno
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Registrar un usuario
      description: |
        Crea un nuevo usuario en Cognito y en DynamoDB de forma atómica.

        **Flujo según rol:**

        - `CUSTOMER_OPERATOR` (defecto): crea el usuario y, si se provee `org_id`,
          escribe el link `ORGANIZATION#<org_id>/OPERATOR#<user_id>`.
        - `CUSTOMER`: crea el usuario y la organización en un `TransactWrite` de 3 ítems:
          `ORGANIZATION#<org_id>/#METADATA` + `USER#<user_id>/#METADATA` +
          `ORGANIZATION#<org_id>/USER#<user_id>`. Requiere `org_name`.
        - `ADMIN`: crea solo el usuario sin organización.

        **Rollback:** si DynamoDB falla después de crear el usuario en Cognito,
        se llama `AdminDeleteUser` automáticamente para no dejar un huérfano.
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            examples:
              operador:
                summary: Operador asignado a una organización
                value:
                  rut: "12345678-5"
                  password: "MiPass123!"
                  email: "juan.perez@empresa.cl"
                  given_name: "Juan"
                  family_name: "Pérez"
                  phone_number: "+56912345678"
                  role: "CUSTOMER_OPERATOR"
                  org_id: "c0c0c0c0-0000-0000-0000-000000000000"
                  location_id: "l1l1l1l1-0000-0000-0000-000000000000"
              cliente_empresa:
                summary: Usuario CUSTOMER + nueva organización (atómico)
                value:
                  rut: "87654321-K"
                  password: "OtraPass456!"
                  email: "admin@miempresa.cl"
                  given_name: "María"
                  family_name: "López"
                  role: "CUSTOMER"
                  org_name: "Estacionamiento Central SpA"
                  org_rut: "76.543.210-K"
      responses:
        "201":
          description: Usuario (y organización si aplica) creados exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
              examples:
                operador:
                  value:
                    message: "usuario_creado"
                    id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                cliente_con_org:
                  value:
                    message: "usuario_creado"
                    id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                    org_id: "c0c0c0c0-0000-0000-0000-000000000000"
        "400":
          description: Datos inválidos o campos requeridos faltantes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                org_name_requerido:
                  value: { error: "org_name_requerido" }
                rut_invalido:
                  value: { error: "rut_invalido" }
        "409":
          description: El RUT ya está registrado en Cognito
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "usuario_ya_existe"
        "500":
          description: Error al crear. Si Cognito OK pero DynamoDB falla se hace rollback.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "error_al_crear_usuario"

  /api/v1/auth/register/batch:
    post:
      tags: [Auth]
      summary: Registro masivo desde Excel
      description: |
        Registra múltiples `CUSTOMER_OPERATOR` a partir de un archivo Excel (`.xlsx`).

        El archivo se envía como campo `file` de un `multipart/form-data`.
        API Gateway HTTP v2 entrega el body en base64 (`IsBase64Encoded: true`).

        **Estructura de la planilla:**

        | Col A | Col B  | Col C    | Col D | Col E    | Col F               |
        |-------|--------|----------|-------|----------|---------------------|
        | rut   | nombre | apellido | email | password | teléfono (opcional) |

        Fila 1 es el header y se ignora. Los datos empiezan en fila 2.

        Los usuarios se procesan en chunks de 25. Los errores son parciales:
        si una fila falla, el resto continúa.
      operationId: registerBatch
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: |
            Procesamiento completado. Revisar `failed` y `errors` — el `200` no
            garantiza que todos los usuarios fueron creados.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchRegisterResponse"
        "400":
          description: Problema con el archivo antes de procesar filas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                archivo_requerido:
                  value: { error: "archivo_requerido" }
                archivo_invalido:
                  value: { error: "archivo_invalido" }

  /api/v1/auth/logout:
    post:
      tags: [Sessions]
      summary: Cerrar sesión propia
      description: |
        Cierra la sesión activa del usuario autenticado:
        1. Llama `GlobalSignOut` en Cognito → invalida el `refresh_token`.
        2. Elimina `USER#<id>/SESSION#ACTIVE` en DynamoDB.

        **Nota:** el `access_token` actual sigue siendo válido hasta su
        expiración natural (~1h). Para bloqueo inmediato en endpoints protegidos
        se requeriría una blocklist de tokens (no implementada en esta versión).
      operationId: logout
      security:
        - CognitoJWT: []
      responses:
        "200":
          description: Sesión cerrada correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
              example:
                message: "sesion_cerrada"
        "401":
          description: Token ausente o inválido (rechazado por API Gateway antes de llegar a la Lambda)
        "404":
          description: Usuario no encontrado en DynamoDB
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "usuario_no_encontrado"
        "500":
          description: Error al eliminar la sesión
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/auth/sessions/{user_id}:
    delete:
      tags: [Sessions]
      summary: Cerrar sesión remota (ADMIN)
      description: |
        Permite a un ADMIN cerrar la sesión activa de cualquier usuario de forma remota.
        Útil cuando el cliente solicita bloquear a un usuario o revocar acceso de emergencia.

        Flujo:
        1. Verifica que el caller pertenezca al grupo `ADMIN` (via `cognito:groups` del JWT).
        2. Obtiene el RUT del usuario de DynamoDB (`USER#<id>/#METADATA`).
        3. Llama `AdminUserGlobalSignOut` en Cognito → invalida todos los refresh tokens.
        4. Elimina `USER#<id>/SESSION#ACTIVE` en DynamoDB.

        **Nota:** igual que en logout propio, el `access_token` actual del usuario
        permanece válido hasta su expiración (~1h).
      operationId: adminCloseSession
      security:
        - CognitoJWT: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID interno del usuario en DynamoDB (campo `id` del UserItem).
      responses:
        "200":
          description: Sesión cerrada correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
              example:
                message: "sesion_cerrada"
        "400":
          description: Parámetro `user_id` ausente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "user_id_requerido"
        "403":
          description: El caller no pertenece al grupo ADMIN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "sin_permiso"
        "404":
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "usuario_no_encontrado"
        "500":
          description: Error al cerrar la sesión
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"